add enhanced data adapter
